+++ strongtalk/src/cpp/main/vm/code/CodeTable.cpp
@@ -28,7 +28,7 @@ CodeTableLink *CodeTable::new_link( NativeMethod *nm, CodeTableLink *n ) {
 void CodeTable::clear() {
 
     //
-    for ( std::int32_t i = 0; i < tableSize; i++ ) {
+    for ( std::size_t i = 0; i < tableSize; i++ ) {
         at( i )->clear();
     }
