+++ strongtalk/src/cpp/main/vm/code/ZoneHeap.hpp
@@ -103,7 +103,7 @@ public:
 
     HeapChunk *get();
 
-    std::int32_t length() const;
+    std::size_t length() const;
 };
